pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "deepgram-sdk>=5.3.0",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "pipecat-ai>=0.0.95",
    "python-dotenv>=1.2.1",
    "websockets>=15.0.1",
    "livekit-agents>=1.3.6",
    "livekit-plugins-liveavatar>=1.3.6",
    "livekit-plugins-openai>=1.3.6",
    "livekit-plugins-deepgram>=1.3.6",
    "livekit-plugins-elevenlabs>=1.3.6",
]
